+++ Integrations/SMIME_Messaging/SMIME_Messaging.py
@@ -17,7 +17,7 @@ def makebuf(text):
 
 
 class Client:
-    def __init__(self, private_key, public_key, recipient_key):
+    def __init__(self, private_key, public_key):
         self.smime = SMIME.SMIME()
 
         public_key_file = NamedTemporaryFile(delete=False)
@@ -30,10 +30,10 @@ class Client:
         self.private_key_file = private_key_file.name
         private_key_file.close()
 
-        recipient_key_file = NamedTemporaryFile(delete=False)
-        recipient_key_file.write(bytes(recipient_key, 'utf-8'))
-        self.recipient_key_file = recipient_key_file.name
-        recipient_key_file.close()
+        # recipient_key_file = NamedTemporaryFile(delete=False)
+        # recipient_key_file.write(bytes(recipient_key, 'utf-8'))
+        # self.recipient_key_file = recipient_key_file.name
+        # recipient_key_file.close()
 
 
 ''' COMMANDS '''
@@ -152,38 +152,46 @@ def decrypt_email_body(client: Client, args: Dict, file_path=None):
     return human_readable, entry_context
 
 
-def sign_and_encrypt(client: Client, args: Dict):
-    """ Sign and encrypt the message
-
-    Args:
-        client: Client
-        args: Dict
-
-    """
+def send_smime(client: Client, args: Dict):
 
     message = args.get('message', '').encode('utf-8')
+    msg_bio = BIO.MemoryBuffer(message)
+    sign = client.private_key_file
+    encrypt = client.public_key_file
 
-    buf = makebuf(message)
-    client.smime.load_key(client.private_key_file, client.public_key_file)
-    p7 = client.smime.sign(buf)
+    if sign:
+        client.smime.load_key(client.private_key_file, client.public_key_file)
+        if encrypt:
+            p7 = client.smime.sign(msg_bio, flags=SMIME.PKCS7_TEXT)
+        else:
+            p7 = client.smime.sign(msg_bio, flags=SMIME.PKCS7_TEXT|SMIME.PKCS7_DETACHED)
+        msg_bio = BIO.MemoryBuffer(message)  # Recreate coz sign() has consumed it.
 
-    x509 = X509.load_cert(client.recipient_key_file)
-    sk = X509.X509_Stack()
-    sk.push(x509)
-    client.smime.set_x509_stack(sk)
+    if encrypt:
+        x509 = X509.load_cert(client.public_key_file)
+        sk = X509.X509_Stack()
+        sk.push(x509)
+        client.smime.set_x509_stack(sk)
 
-    client.smime.set_cipher(SMIME.Cipher('des_ede3_cbc'))
-
-    tmp = BIO.MemoryBuffer()
-
-    client.smime.write(tmp, p7)
-
-    p7 = client.smime.encrypt(tmp)
+        client.smime.set_cipher(SMIME.Cipher('des_ede3_cbc'))
+        tmp_bio = BIO.MemoryBuffer()
+        if sign:
+            client.smime.write(tmp_bio, p7)
+        else:
+            tmp_bio.write(message)
+        p7 = client.smime.encrypt(tmp_bio)
 
     out = BIO.MemoryBuffer()
-    client.smime.write(out, p7)
-    msg = out.read().decode('utf-8')
+    if encrypt:
+        client.smime.write(out, p7)
+    else:
+        if sign:
+            client.smime.write(out, p7, msg_bio, SMIME.PKCS7_TEXT)
+        else:
+            out.write('\r\n')
+            out.write(message)
 
+    msg = out.read().decode('utf-8')
     entry_context = {
         'SMIME.Encrypted': {
             'Message': msg
@@ -214,9 +222,9 @@ def main():
 
     public_key: str = demisto.params().get('public_key', '')
     private_key: str = demisto.params().get('private_key', '')
-    recipient_key: str = demisto.params().get('recipient_key', '')
+    # recipient_key: str = demisto.params().get('recipient_key', '')
 
-    client = Client(private_key, public_key, recipient_key)
+    client = Client(private_key, public_key)
     LOG(f'Command being called is {demisto.command()}')
     commands = {
         'test-module': test_module,
@@ -224,7 +232,7 @@ def main():
         'smime-encrypt-email-body': encrypt_email_body,
         'smime-verify-sign': verify,
         'smime-decrypt-email-body': decrypt_email_body,
-        'smime-sign-and-encrypt': sign_and_encrypt
+        'smime-send-smime': send_smime
     }
     try:
         command = demisto.command()
@@ -239,8 +247,8 @@ def main():
             os.unlink(client.private_key_file)
         if client.public_key_file:
             os.unlink(client.public_key_file)
-        if client.recipient_key_file:
-            os.unlink(client.recipient_key_file)
+        # if client.recipient_key_file:
+        #     os.unlink(client.recipient_key_file)
 
 
 if __name__ in ['__main__', 'builtin', 'builtins']: